refactor(plugin-swarm): rename callCollaborator tool file and update import
